refactor: improve utils structure
