+++ src/nile/core/commands/clean.py
