+++ src/nile/core/commands/declare.py
