+++ src/nile/core/types/account.py
@@ -13,7 +13,8 @@ from nile.common import (
     is_alias,
     normalize_number,
 )
-from nile.core.deploy import deploy_account
+from nile.core.commands.deploy import deploy_account
+from nile.core.commands.getters.get_nonce import get_nonce_without_log as get_nonce
 from nile.core.types.transactions import DeclareTransaction, InvokeTransaction
 from nile.core.types.tx_wrappers import (
     DeclareTxWrapper,
@@ -26,7 +27,6 @@ from nile.core.types.utils import (
     get_deploy_account_hash,
     get_execute_calldata,
 )
-from nile.utils.get_nonce import get_nonce_without_log as get_nonce
 
 try:
     from nile.core.types.signer import Signer
