+++ tests/commands/test_get_balance.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 import pytest
 
 from nile.common import ETH_TOKEN_ABI, ETH_TOKEN_ADDRESS
-from nile.utils.get_balance import get_balance
+from nile.core.commands.getters.get_balance import get_balance
 
 NETWORKS = ["mainnet", "goerli", "goerli2", "localhost"]
 CONTRACTS = ["0x4d2", "1234", 1234]
@@ -21,7 +21,8 @@ EXPECTED_VALUES = [
 @pytest.mark.parametrize("mock_return", EXPECTED_VALUES)
 async def test_get_balance(contract_address, network, mock_return):
     with patch(
-        "nile.utils.get_balance.call_or_invoke", return_value=mock_return[0]
+        "nile.core.commands.getters.get_balance.call_or_invoke",
+        return_value=mock_return[0],
     ) as mock_call:
         res = await get_balance(contract_address, network)
